requirements-dev.txt
tox<3
virtualenv<16
pytest
sphinx
sphinxcontrib-programoutput
sphinx_rtd_theme
pyyaml
sh
